backend/requirements.txt
amqp==5.3.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.25
django-cors-headers==4.1.0
django-redis==5.4.0
idna==3.10
importlib-metadata==4.12.0
kombu==5.2.4
prompt_toolkit==3.0.48
python-dotenv==0.21.1
pytz==2024.2
redis==5.0.8
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
zipp==3.15.0
